Moving all express routing into its own file
